% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyPrepare.R
\name{classifyPrepare}
\alias{classifyPrepare}
\title{Prepare classification data}
\usage{
classifyPrepare(
  data,
  label = character(),
  discrete = list(),
  continuous = list(),
  ratios = list()
)
}
\arguments{
\item{data}{a \code{data.frame} object containing observations.}

\item{label}{a character indicating the column in \code{data} to use as the
classification outcome.}

\item{discrete}{a character vector of columns representing discrete
variables.}

\item{continuous}{a character vector of columns representing continuous
variables.}

\item{ratios}{a character vector of columns representing ratio variables.}
}
\value{
An \code{\link{ELSIPData}} object.
}
\description{
This function processes datasets containing numerical and/or discrete value
datasets into a format suitable for analysis using \code{\link{ELSIP}} using
transformations specifically suited to each type of data. Presently, three
types of data types are supported by \code{classifyPrepare}: continuous,
ratio, and discrete. The variables desired are specified according to the
named columns in \code{data} for any of the three aforementioned data types.
Named columns in \code{discrete}, \code{continuous} and \code{ratios} will
appear in the returned data frame, with the following transformations
applied:
\itemize{
  \item{\code{discrete} variables are included as is, but converted to
  \code{factor}s;}
  \item{\code{continuous} variables are log-10 transformed; and}
  \item{\code{ratios} variables are logit transformed. Note that the upper
  and lower bounds of the logit transformation are set to minus and plus 1
  of the lowest and highest value of the ratio variable, respectively, to
  avoid returning \code{Inf}s.}
}
}
