% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELSIPData.R
\name{ELSIPData}
\alias{ELSIPData}
\title{ELSIPData class}
\description{
ELSIPData class
}
\section{Fields}{

\describe{
\item{\code{x}}{a data frame object where samples are in rows and features are in
columns.}

\item{\code{y}}{a factor vector indicating the observed class for each row. All
samples intended for model training and testing purposes should have an
assigned class. Any sample that does not have an assigned class
(i.e., \code{NA}) will be interpreted as an unknown sample and classified
after model training and testing.}

\item{\code{data_type}}{an optional factor vector indicating the model class for
each observation. The allowed options are \code{"Training"},
\code{"Test"}, and \code{"Unknown"}.}
}}

