% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELSIPTrain.R
\name{ELSIPTrain}
\alias{ELSIPTrain}
\title{ELSIPTrain class}
\description{
ELSIPTrain class

ELSIPTrain class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model}}{A \code{train} object generated by
\code{\link[caret]{train}} list containing the trained classification
model. This is a read only value.}

\item{\code{data}}{The \code{\link{ELSIPData}} object containing the training
data used to generate \code{model}. This is a read only value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ELSIPTrain$new()}}
\item \href{#method-print}{\code{ELSIPTrain$print()}}
\item \href{#method-clone}{\code{ELSIPTrain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{ELSIPTrain} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ELSIPTrain$new(model, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A \code{train} object generated by
\code{\link[caret]{train}}.}

\item{\code{data}}{An \code{\link{ELSIPData}} object. The training partition
from this object must have the same dimensions, column names, values
and outcome variables as the training data in \code{train}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{ELSIPTrain} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print an \code{ELSIPData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ELSIPTrain$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ELSIPTrain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
