% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Ensemble ELSIP predictions}
\usage{
ensemble(..., classifier, pairwise_stack = FALSE, ensemble_single = FALSE)
}
\arguments{
\item{...}{One or more objects of class \code{\link{ELSIPTrain}} created
from \code{\link[ELSIP:train.ELSIPData]{train}}.}

\item{classifier}{a list defining the model and model parameters to use
during classification of the ensembled datasets. The list should contain
the following named items:
\describe{
  \item{method}{a single length character vector indicating the
    classification method to use with \code{caret}.}
  \item{trainControl}{a list generated by
    \code{\link[caret:trainControl]{trainControl}}.}
  \item{tuneGrid}{a list defining the tuning grid to use for
    training the \code{caret} model.}
}}

\item{pairwise_stack}{a logical value indicating whether to stack
predictions from only from the training/testing set the classifiers were
trained on (\code{FALSE}; default) or all pairwise combinations of
training/testing imputations and classifiers (\code{TRUE}).}

\item{ensemble_single}{a logical value indicating whether to perform
ensembling on a singly imputed dataset (i.e., when \code{imp_n} is 1).}
}
\description{
Ensemble ELSIP predictions
}
