% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute.ELSIPData}
\alias{impute.ELSIPData}
\title{Impute multivariate data}
\usage{
\method{impute}{ELSIPData}(x, method = "mean", n = 1, seed = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{object that inherits \code{\link{ELSIPData}}, such as the output
of \code{\link{classifyPrepare}}.}

\item{method}{a single length character indicating the imputation method to
use. Options are "\code{mean}" (default), "\code{median}",
"\code{mice_pmm}", "\code{mice_rf}" and "\code{mipca}".}

\item{n}{a single length numeric or integer greater than 0 indicating the
number of imputations to generate. Note that multiple imputations (i.e.,
values of \code{n} greater than 1, cannot be generated when \code{method}
is "\code{mean}" or "\code{median}".}

\item{seed}{an integer value to initialise the multiple imputation.}

\item{verbose}{a logical indicating wether to output information to the
console during multiple imputation.}

\item{...}{ignored}
}
\description{
Impute multivariate data
}
