% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiELSIPTrain.R
\name{multiELSIPTrain}
\alias{multiELSIPTrain}
\title{multiELSIPTrain class}
\description{
multiELSIPTrain class

multiELSIPTrain class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{train}}{A list of two or more objects of type
\code{\link{ELSIPTrain}}. This is a read only value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{multiELSIPTrain$new()}}
\item \href{#method-print}{\code{multiELSIPTrain$print()}}
\item \href{#method-clone}{\code{multiELSIPTrain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{multiELSIPTrain} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiELSIPTrain$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more \code{\link{ELSIPTrain}} objects. The training
data of each object must contain the same number of rows and identical
outcome variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{multiELSIPTrain} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{multiELSIPTrain} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiELSIPTrain$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multiELSIPTrain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
