% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition multivariate data}
\usage{
partition(data, train_prop, seed = NULL)
}
\arguments{
\item{data}{object that inherits \code{\link{ELSIPData}}, such as the output
of \code{\link{classifyPrepare}}.}

\item{train_prop}{a numeric indicating what proportion of observations will
constitute the training dataset. Must be a number between 0 and 1, not
inclusive.}

\item{seed}{a random seed for initialising the creation of partitions.}
}
\description{
Partition multivariate data
}
