% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.ELSIPTrain.R
\name{print.summary.ELSIPTrain}
\alias{print.summary.ELSIPTrain}
\title{Summarize ELSIPTrain object}
\usage{
\method{print}{summary.ELSIPTrain}(x, ...)
}
\arguments{
\item{x}{An \code{\link{ELSIPTrain}} object.}

\item{...}{Optional arguments (currently none.)}
}
\description{
Summarize ELSIPTrain object
}
