% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsample multivariate datasets}
\usage{
subsample(data, type = c("up", "down"), seed = NULL)
}
\arguments{
\item{data}{an \code{\link{ELSIPData}} object, such as the output of
\code{\link{classifyPrepare}}.}

\item{type}{a single length character indicating the type of subsampling to
perform. Current options are \code{"up"} (upsampling; using
\code{\link[caret]{upSample}}) and \code{"down"}
(downsampling; using \code{\link[caret]{downSample}}).}

\item{seed}{a random seed for initializing the subsampling.}
}
\value{
An \code{\link{ELSIPData}} object.
}
\description{
This function performs subsampling of multivariate datasets where the
proportions of the outcome variables may be inbalanced, which can lead to
poor preduction accuracy.
}
