% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ELSIPTrain.R
\name{summary.ELSIPTrain}
\alias{summary.ELSIPTrain}
\title{Summarize ELSIPTrain object}
\usage{
\method{summary}{ELSIPTrain}(object, ...)
}
\arguments{
\item{object}{An \code{\link{ELSIPTrain}} object.}

\item{...}{Optional arguments (currently none.)}
}
\description{
Summarize ELSIPTrain object
}
