% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.ELSIP.R
\name{train.multiELSIPData}
\alias{train.multiELSIPData}
\title{ELSIP trainer}
\usage{
\method{train}{multiELSIPData}(data, classifier)
}
\arguments{
\item{data}{An \code{\link{multiELSIPData}} object generated from
\code{\link[ELSIP]{impute}}.}

\item{classifier}{a list defining the model and model parameters to use
during classification of the multiply imputed training and testing data.
The list should contain the following named items:
\describe{
  \item{method}{a single length character vector indicating the
    classification method to use with \code{caret}.}
  \item{trainControl}{a list generated by
    \code{\link[caret:trainControl]{trainControl}}.}
  \item{tuneGrid}{a list defining the tuning grid to use for
    training the \code{caret} model.}
}}
}
\description{
ELSIP trainer
}
