% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELSIP.R
\name{ELSIP}
\alias{ELSIP}
\title{Ensemble learning using stacked imputation predictions}
\usage{
ELSIP(
  data,
  subsample = c("none", "up", "down"),
  n_imp = 10,
  imp_method = c("mean", "median", "mice_pmm", "mice_rf", "mipca"),
  train_prop = logical(),
  classifier = list(),
  pairwise_stack = TRUE,
  meta_classifier = list(),
  ensemble_single = FALSE,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{An \code{\link{ELSIPData}} object.}

\item{subsample}{a single length character indicating which
subsampling to apply to the training set prior to model tuning. The
default (\code{"none"}) is for no subsampling; options are \code{"up"}
and \code{"down"}.}

\item{n_imp}{an integer indicating the number of multiple imputations to
generate.}

\item{imp_method}{a character vector indicating the imputation method to
use with \code{mice}. Can be either a single character to use for all
variables, or a vector indicating the method to use for each variable.}

\item{train_prop}{a numeric indicating what proportion of observations will
constitute the training dataset.}

\item{classifier}{a list defining the model and model parameters to use
during classification of the multiply imputed datasets. The list should
contain the following named items:
\describe{
  \item{method}{a single length character vector indicating the
    classification method to use with \code{caret}.}
  \item{trainControl}{a list generated by
    \code{\link[caret:trainControl]{trainControl}}.}
  \item{tuneGrid}{a list defining the tuning grid to use for
    training the \code{caret} model.}
}}

\item{pairwise_stack}{a logical value indicating whether to stack
predictions from only from the training/testing set the classifiers were
trained on (\code{FALSE}; default) or all pairwise combinations of
training/testing imputations and classifiers (\code{TRUE}).}

\item{meta_classifier}{a list defining the model and model parameters to use
during classification of the ensemble classification. Structure follows
that of \code{classifier}.}

\item{ensemble_single}{a logical value indicating whether to perform
ensembling on a singly imputed dataset (i.e., when \code{imp_n} is 1).}

\item{seed}{an integer value to initialise the multiple imputation.}

\item{verbose}{output information to the console during multiple imputation.}
}
\value{
An \code{\link{ELSIPTrain}} object.
}
\description{
This is a convenience function for performing ensemble learning in
\code{ELSIP}. It equivalent to using, in order, \code{\link{subsample}}
(optional), \code{\link{partition}}, \code{\link{impute}},
\code{\link{train.ELSIPData}}, and \code{\link{ensemble}} (optional).
}
