% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELSIPTrain.R
\name{ELSIPTrain}
\alias{ELSIPTrain}
\title{ELSIPTrain class}
\description{
ELSIPTrain class
}
\section{Fields}{

\describe{
\item{\code{model}}{A \code{train} object generated by \code{\link[caret]{train}}
list containing the trained classification model.}

\item{\code{data}}{The \code{\link{ELSIPData}} object used to generate the
training model in \code{model}.}
}}

