% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiELSIPData.R
\name{multiELSIPData}
\alias{multiELSIPData}
\title{multiELSIPData class}
\description{
multiELSIPData class
}
\section{Fields}{

\describe{
\item{\code{data}}{Two or more objects of type \code{\link{ELSIPData}}.}
}}

