% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiELSIPTrain.R
\name{multiELSIPTrain}
\alias{multiELSIPTrain}
\title{multiELSIPTrain class}
\description{
multiELSIPTrain class
}
\section{Fields}{

\describe{
\item{\code{.model}}{A list containing the results of
\code{\link{train.ELSIPData}}. The following named items are included:
\describe{
  \item{label}{The full name of the classification model.}
  \item{library}{The name of the R library implementing the classification
    model.}
  \item{parameters}{The list of parameters specified with the
    classification model.}
  \item{method}{The name of the classification model as is appears in
    \code{\link[caret]{getModelInfo}}.}
  \item{res}{A list of objects of type \code{train} generated by
    \code{\link[caret]{train}} containing the results of the
    classification.}
}}
}}

