% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition multivariate data}
\usage{
partition(data, train_prop, seed = NULL)
}
\arguments{
\item{data}{an \code{\link{ELSIPData}} object, such as the output of
\code{\link{classifyPrepare}}.}

\item{train_prop}{a numeric indicating what proportion of observations will
constitute the training dataset. Must be a number between 0 and 1, not
inclusive.}

\item{seed}{a random seed for initialising the creation of partitions.}
}
\value{
an \code{\link{ELSIPData}} object.
}
\description{
Thus function allocates samples of a dataset into training and testing
samples depending on class. Samples with a \code{NA} class are assigned
to a separate "Unknown" partition.
}
