% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsample multivariate datasets}
\usage{
subsample(data, type = c("up", "down"), seed = NULL)
}
\arguments{
\item{data}{an object that inherits \code{\link{ELSIPData}}, such as the output
of \code{\link{classifyPrepare}}.}

\item{type}{a single length character indicating the type of subsampling to
perform.}

\item{seed}{a random seed for initializing the subsampling.}
}
\value{
An \code{\link{ELSIPData}} object.
}
\description{
Subsample multivariate datasets
}
